hzgml/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
atlas-mpl-style==0.20.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
cycler==0.10.0
dataclasses==0.6
fastparquet==0.8.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.1
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
numexpr==2.7.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
root-numpy==4.8.0
root-pandas==0.7.0
scikit-learn==0.24.2
scikit-opt==0.6.5
scipy==1.5.4
six==1.15.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-addons==0.14.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.0.0
tqdm==4.62.3
typeguard==2.13.0
typing-extensions==3.7.4.3
uproot==4.1.5
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
xgboost==0.90
zipp==3.6.0
